use size_t for getting a slice. clang-format induced reformatting

+++ src/partition.h
@@ -23,79 +23,56 @@
 
 #include <vector>
 
-#include "types.h"
 #include "common.h"
 #include "machine.h"
+#include "types.h"
 
 namespace tomocam {
 
     template <typename T>
     class Partition {
-        private:
-            dim3_t dims_;
-            uint64_t size_;
-            T *first_;
-            int halo_[2];
+      private:
+        dim3_t dims_;
+        uint64_t size_;
+        T *first_;
+        int halo_[2];
 
-        public:
-            Partition(dim3_t d, T *pos) : dims_(d), first_(pos) {
-                size_ = static_cast<uint64_t>(dims_.z) * dims_.y * dims_.x;
-                halo_[0] = 0;
-                halo_[1] = 0;
-            }
+      public:
+        Partition(dim3_t d, T *pos) : dims_(d), first_(pos) {
+            size_ = static_cast<uint64_t>(dims_.z) * dims_.y * dims_.x;
+            halo_[0] = 0;
+            halo_[1] = 0;
+        }
 
-            Partition(dim3_t d, T *pos, int *h) : dims_(d), first_(pos) {
-                size_ = static_cast<uint64_t>(dims_.z) * dims_.y * dims_.x;
-                halo_[0] = h[0];
-                halo_[1] = h[1];
-            }
+        Partition(dim3_t d, T *pos, int *h) : dims_(d), first_(pos) {
+            size_ = static_cast<uint64_t>(dims_.z) * dims_.y * dims_.x;
+            halo_[0] = h[0];
+            halo_[1] = h[1];
+        }
 
-            dim3_t dims() const {
-                return dims_;
-            }
-            int nslices() const {
-                return dims_.x;
-            }
-            int nrows() const {
-                return dims_.y;
-            }
-            int ncols() const {
-                return dims_.z;
-            }
-            uint64_t size() const {
-                return size_;
-            }
-            size_t bytes() const {
-                return size_ * sizeof(T);
-            }
-            int  *halo() {
-                return halo_;
-            }
-            const int  *halo() const {
-                return halo_;
-            }
+        dim3_t dims() const { return dims_; }
+        int nslices() const { return dims_.x; }
+        int nrows() const { return dims_.y; }
+        int ncols() const { return dims_.z; }
+        uint64_t size() const { return size_; }
+        size_t bytes() const { return size_ * sizeof(T); }
+        int *halo() { return halo_; }
+        const int *halo() const { return halo_; }
 
-            T *begin() {
-                return first_;
-            }
-            const T *begin() const {
-                return first_;
-            }
-            T *slice(int i) {
-                return first_ + i * dims_.y * dims_.z;
-            }
-            const T *slice(int i) const {
-                return first_ + i * dims_.y * dims_.z;
-            }
+        T *begin() { return first_; }
+        const T *begin() const { return first_; }
+        T *slice(size_t i) { return first_ + i * dims_.y * dims_.z; }
+        const T *slice(size_t i) const { return first_ + i * dims_.y * dims_.z; }
 
-            T &operator()(int i, int j, int k) {
-                return first_[i * dims_.y * dims_.z + j * dims_.z + k];
-            }
-            T operator()(int i, int j, int k) const {
-                return first_[i * dims_.y * dims_.z + j * dims_.z + k];
-            }
+        T &operator()(int i, int j, int k) {
+            return first_[static_cast<size_t>(i) * dims_.y * dims_.z +
+                          static_cast<size_t>(j) * dims_.z + k];
+        }
+        T operator()(int i, int j, int k) const {
+            return first_[static_cast<size_t>(i) * dims_.y * dims_.z +
+                          static_cast<size_t>(j) * dims_.z + k];
+        }
     };
-
 
     inline std::vector<std::vector<int>> create_partition_table(int N) {
 
@@ -106,21 +83,15 @@ namespace tomocam {
         int n_p = N / nBlocks;
         int n_e = N % nBlocks;
         std::vector<int> pre_partition;
-        for (int i = 0; i < n_p; i++) {
-            pre_partition.push_back(i * nBlocks);
-        }
-        if (n_e > 0) {
-            pre_partition.push_back(n_p * nBlocks);
-        }
-        
+        for (int i = 0; i < n_p; i++) { pre_partition.push_back(i * nBlocks); }
+        if (n_e > 0) { pre_partition.push_back(n_p * nBlocks); }
+
         int nParts = pre_partition.size() / nDevices;
         int nExtra = pre_partition.size() % nDevices;
-       
+
         // create the partition table
         std::vector<int> sizes(nDevices, nParts);
-        for (int i = 0; i < nExtra; i++) {
-            sizes[i] += 1;
-        }
+        for (int i = 0; i < nExtra; i++) { sizes[i] += 1; }
 
         int ibegin = 0;
         std::vector<std::vector<int>> table(nDevices);
@@ -133,12 +104,10 @@ namespace tomocam {
         }
         return table;
     }
-     
-    
+
     // partition an array into sub-partitions
     template <typename T>
-    std::vector<Partition<T>> create_partitions(Partition<T> &a,
-        int npartitions) {
+    std::vector<Partition<T>> create_partitions(Partition<T> &a, int npartitions) {
 
         // create sub-partitions
         std::vector<Partition<T>> table;
@@ -158,8 +127,8 @@ namespace tomocam {
 
     // partition an array into sub-partitions with halo in x-direction
     template <typename T>
-    std::vector<Partition<T>> create_partitions(Partition<T> &a,
-        int npartitions, int halo) {
+    std::vector<Partition<T>> create_partitions(Partition<T> &a, int npartitions,
+                                                int halo) {
 
         // sanity check
         if (npartitions < 0) {
@@ -189,17 +158,19 @@ namespace tomocam {
 
         // set pointers as if  there is no halo
         std::vector<int> shares(npartitions, work);
-        for (int i = 0; i < extra; i++) {
-            shares[i] += 1;
-        }
+        for (int i = 0; i < extra; i++) { shares[i] += 1; }
 
         // create the sub-partitions
         int h[2];
         for (int i = 0; i < npartitions; i++) {
-            if (i == 0) h[0] = a.halo()[0];
-            else h[0] = halo;
-            if (i == npartitions - 1) h[1] = a.halo()[1];
-            else h[1] = halo;
+            if (i == 0)
+                h[0] = a.halo()[0];
+            else
+                h[0] = halo;
+            if (i == npartitions - 1)
+                h[1] = a.halo()[1];
+            else
+                h[1] = halo;
             dim3_t d(shares[i] + h[0] + h[1], dims.y, dims.z);
             table.push_back(Partition<T>(d, a.slice(offset - h[0]), h));
             offset += shares[i];
